flask-api/requirements.txt
Flask
flask-cors
pillow
numpy
opencv-python-headless
scikit-image
pytesseract
gunicorn